refactor(desktop): split workspace file editor components
